requirements.txt
cryptography
pillow
qrcode
md5
AES
pyzbar